Fix dto query fields

+++ final-api/src/main/java/com/salesianostriana/dam/finalapi/dtos/site/GetListSiteDto.java
@@ -16,7 +16,7 @@ public class GetListSiteDto {
     private String address;
     private String city;
     private String postalCode;
-    private int totalComments;
+    private long totalComments;
     private double rate;
     private String scaledFileUrl;
     private boolean liked;
